Some more fixes to the timing calculations.
If the vsync is enabled missuse it to slow the client down, e.g.
calculate the target framerate, add an security margin of 20% and
let the vsync handle the rest. This hopefully solves some problems
with frametime spikes. This is an idea by @DanielGibson.

If the vsync is disabled use a somewhat more conservative average
frametime calculation. Take only pure render- and paketframes into
account, but not combined frames. And only the last 30 instead of
60 frames.
It's a little bit questionable if the average frametimes are really
necessary, on a rather fast Geforce and Core i7 system I don't see
any differences with the average frametimes taken into account or not.
Maybe we should make them optional.